% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparat-package.R
\docType{package}
\name{comparat-package}
\alias{comparat}
\alias{comparat-package}
\title{comparat: comparat}
\description{
Comparat provides a Shiny application for comparing means and variances for two indipendent samples.
}
\author{
\strong{Maintainer}: Andrea Bazzano \email{andrea.bazzano@arpal.liguria.it}

}
\keyword{internal}
