% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compare.R
\name{fct_gesd}
\alias{fct_gesd}
\title{Displays the results of generelised extreme strudentized deviate (GESD) test}
\source{
UNI ISO 16269-4:2019 - Statistical interpretation of data -
 Part 4: Detection and treatment of outliers. Section 4.3.2 and Annex A.
 \url{https://store.uni.com/uni-iso-16269-4-2019}
}
\usage{
fct_gesd(values, significance = 0.95, m = round(length(values)/3, 0))
}
\arguments{
\item{values}{a \code{vector} with the values relevant for testing.}

\item{significance}{a number, typically either 0.90, 0.95 (default) or 0.99
indicating the confidence level for the test.}

\item{m}{maximum number of possible outliers to be tested.
Default is one third of the number of values.}
}
\value{
A dataframe  with the following columns:
\describe{
   \item{I}{Numeric values inspected by the test.}
   \item{R}{Numeric values for the extreme studentized deviate.}
   \item{lambda}{numberic values with the critical values of the test statistics.}
   \item{outliers}{a logical vector with the result of the test.}
 }
}
\description{
The function displays the results of generelised
 extreme strudentized deviate (GESD) test for outlier detection.
 The returned text is suitable for the {comparat} {shiny} app.
}
\details{
the GESD test is performed according to secton 4.3.2 of
 UNI ISO 16269-4:2019 - Statistical interpretation of data - Part 4: Detection
 and treatment of outliers. The software implementation in performed in
 accordance to Annex A of the same document.
}
