% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compare.R
\name{fct_grubbs}
\alias{fct_grubbs}
\title{Displays the results of Grubbs tests for outlier detection}
\usage{
fct_grubbs(x, significance = 0.95)
}
\arguments{
\item{significance}{a number, typically either 0.90, 0.95 (default) or 0.99
indicating the confidence level for the test.}

\item{values}{a \code{vector} with the values relevant for testing.}
}
\value{
A string with the result of the test.
}
\description{
The function displays the results of repeated Grubbs test for
 outlier detection.
 The returned text is suitable for the {comparat} {shiny} app.
}
\details{
the Grubbs test is calculated using the function \code{grubbs.test}
provided by the \code{outliers} package.
}
