% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{rowsummary}
\alias{rowsummary}
\title{Summary arranged on rows for two or more groups}
\usage{
rowsummary(data, response, group, signif = 3L)
}
\arguments{
\item{data}{the \code{data.frame} or \code{data.table} to be summarised.}

\item{response}{a string with the name of the variable to summarise.}

\item{group}{a string with the name of the grouping variable.}

\item{signif}{a integer with the number of desired significant figures.}
}
\value{
a \code{data.table} with 6 rows and \eqn{n + 1} columns for \eqn{n}
  levels of the group variable.
}
\description{
The function returns a table with max, mean, median, min, sd and n
 values arranged on rows while groups are on columns. Numbers are formatted as
 text in order to provide the desired significant figures.
}
