% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compare.R
\name{fct_grubbs}
\alias{fct_grubbs}
\title{Displays the results of Grubbs tests for outlier detection}
\usage{
fct_grubbs(values, significance = 0.95)
}
\arguments{
\item{values}{a \code{vector} with the values relevant for testing.}

\item{significance}{a number, typically either 0.90, 0.95 (default) or 0.99
indicating the confidence level for the test.}
}
\value{
A list  with the following items:
\describe{
   \item{G}{a numeric value with the test statistic.}
   \item{pvalue}{a numeric value with the p-value of the test.}
   \item{result}{A string with the result of the test.}
 }
}
\description{
The function displays the results of repeated Grubbs test for
 outlier detection.
 The returned text is suitable for the {comparat} {shiny} app.
}
\details{
the Grubbs test is calculated using the function \code{grubbs.test}
provided by the \code{outliers} package. The function checks whether
a single or two values, both at the same or opposite sides of the remaining
values can be considered outliers.
}
