% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIconfronta-package.R
\docType{package}
\name{SIconfronta-package}
\alias{SIconfronta}
\alias{SIconfronta-package}
\title{SIconfronta: SIconfronta}
\description{
Comparat provides a Shiny application for comparing means and variances for two indipendent samples.
}
\author{
\strong{Maintainer}: Andrea Bazzano \email{andrea.bazzano@arpal.liguria.it}

}
\keyword{internal}
